Tests: Nicer error exit for compile extension modules test.
* Also the test was not working for Python3 yet.

* Also avoid warnings that are disabled with PyLint annotations.

+++ tests/library/compile_extension_modules.py
@@ -19,7 +19,7 @@
 #     limitations under the License.
 #
 
-import os, sys, tempfile, subprocess, shutil
+import os, sys, tempfile, shutil
 
 # Find nuitka package relative to us.
 sys.path.insert(
@@ -83,7 +83,8 @@ def action(stage_dir, root, path):
         "--run",
         "--output-dir",
         stage_dir,
-        "--remove-output"
+        "--remove-output",
+        "--plugin-enable=pylint-warnings"
     ]
 
     filename = os.path.join(stage_dir, "importer.py")
@@ -104,11 +105,11 @@ def action(stage_dir, root, path):
 
     try:
         output = check_output(command).splitlines()
-        assert output[-1] == "OK", output
-    except Exception as e:
-        print(e)
-        sys.exit(1)
+    except Exception:
+        raise
     else:
+        if output[-1] != b"OK":
+            sys.exit("FAIL")
         my_print("OK")
 
         shutil.rmtree(filename[:-3] + ".dist")
